requirements.txt
certifi==2024.2.2
clickhouse-connect==0.7.3
lz4==4.3.3
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
zstandard==0.22.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0